% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate_Result_List.R
\name{Aggregate.Result.List}
\alias{Aggregate.Result.List}
\title{Aggregate result list}
\usage{
Aggregate.Result.List(result_list, id, complex = F)
}
\arguments{
\item{result_list}{[\code{list(1)}]\cr Result list create with Result.Habitat.Association()}

\item{id}{[\code{string(1)}]\cr Name of the ID-variable}

\item{complex}{[\code{logical(1)}] \cr TRUE if result_list is based on complex simulation pattern}
}
\value{
Long data frame with aggregated results
}
\description{
Aggregates the result of FUNCTION-NAME into the long format and calculates some statistics (e.g. mean, SE, ...)
}
