% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_result_list.R
\name{aggregate_result_list}
\alias{aggregate_result_list}
\title{Aggregate result list}
\usage{
aggregate_result_list(result_list, id, neutral = F)
}
\arguments{
\item{result_list}{[\code{list(1)}]\cr Result list create with Result.Habitat.Association()}

\item{id}{[\code{string(1)}]\cr Name of the ID-variable}

\item{neutral}{[\code{logical(1)}] \cr TRUE if result_list is based on neutral species}
}
\value{
Long data frame with aggregated results
}
\description{
Aggregates the result of FUNCTION-NAME into the long format and calculates some statistics (e.g. mean, SE, ...)
}
