% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gest.R
\name{calc_gest}
\alias{calc_gest}
\title{calc_gest}
\usage{
calc_gest(dist, r, n_points)
}
\arguments{
\item{dist}{matrix with distance pairs.}

\item{r}{vector with distances r.}

\item{n_points}{numeric with number of points}
}
\value{
data.frame
}
\description{
Calculate Gest
}
\details{
Calculates Gest based on distances created with \code{get_dist_pairs}.
}
\seealso{
\code{\link{get_dist_pairs}}
}
\keyword{internal}
