% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_moments.R
\name{calc_moments}
\alias{calc_moments}
\title{calc_moments}
\usage{
calc_moments(fn, p, exclude = NULL, x, y, mark, kernel, rmax_bw, r)
}
\arguments{
\item{fn}{Determination of the weightings of the mark correlation functions.}

\item{p}{Defines the initial state of the new ponit pattern.}

\item{exclude}{}

\item{x}{x Coordinate of the points from the reference point pattern.}

\item{y}{y Coordinate of the points from the reference point pattern.}

\item{mark}{Marks the currently viewed point pattern.}

\item{kernel}{Result of the kernel calculation, calculated with the
calc_kernels function.}

\item{rmax_bw}{Maximum distance at which the summary statistics are
evaluated + Bandwidth with which the kernels are scaled, so that this is the
 standard deviation of the smoothing kernel.}

\item{r}{is a sequence from rmin to rmax in rcount steps.}
}
\value{
matrix
}
\description{
calc_moments (internal)
}
\details{
Definition of the product-moment function for calculating the contribution
of a point at the coordinates x, y with marking.
}
\keyword{internal}
