% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_energy.R
\name{calculate_mean_energy}
\alias{calculate_mean_energy}
\title{Mean energy of the reconstruction}
\usage{
calculate_mean_energy(pattern, only_spatial = TRUE)
}
\arguments{
\item{pattern}{[\code{list(1)}] List with reconstructed and observed point pattern created with SHAR::Pattern.Reconstruction()}

\item{only_spatial}{[\code{logical(1)}] TRUE for univariate point pattern}
}
\value{
Numeric with mean energy of the reconstruction
}
\description{
Function to plot all reconstructed patterns created of the pattern list with reconstructed patterns
}
