% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classint_to_vector.R
\name{classint_to_vector}
\alias{classint_to_vector}
\title{classint_to_vector}
\usage{
classint_to_vector(x, digits = NULL)
}
\arguments{
\item{x}{classIntervals object}

\item{digits}{Integer with digits used during rounding.}
}
\value{
vector
}
\description{
Convert classIntervals to vector
}
\details{
Returns a character vector with breaks of a \code{classIntervals} object. If
\code(digits = NULL), results wiill not be rounded
}
\examples{
\dontrun{
classint_to_vector(x = landscape_classified$breaks, digits = 4)
}

}
\keyword{internal}
