% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complex_pattern.R
\name{Complex.Pattern}
\alias{Complex.Pattern}
\title{Help function}
\usage{
\method{Complex}{Pattern}(owin)
}
\description{
Internal help function to create complex pattern
#' @param owin [\code{owin(1)}] Window of the spatstat package
}
\details{
#' @return ppp object of the spatstat package
}
