% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_statistics.R
\name{compute_statistics}
\alias{compute_statistics}
\title{compute_statistics}
\usage{
compute_statistics(x, y, k, xr, yr, w_statistics)
}
\arguments{
\item{x}{x Coordinate of the points from the reference point pattern.}

\item{y}{y Coordinate of the points from the reference point pattern.}

\item{k}{Vector of values k; used only if Dk is included in w_statistics
elow.}

\item{xr}{x extension of the observation window (start, end)}

\item{yr}{y extension of the observation window (start, end)}

\item{w_statistics}{}
}
\value{
list
}
\description{
compute_statistics (internal)
}
\details{
Compute optional spatial statistics using the spatstat package.
}
\keyword{internal}
