% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Simulation_Pattern.R
\name{Create.Simulation.Pattern}
\alias{Create.Simulation.Pattern}
\title{Create simulation pattern}
\usage{
Create.Simulation.Pattern(habitats, number_points = 100, alpha = 0.3,
  complex = F)
}
\arguments{
\item{habitats}{[\code{raster(1)}]\cr Raster object of the raster package with habitats}

\item{number_points}{[\code{numeric(1)}]\cr Number of points for each species (alpha=0)}

\item{alpha}{[\code{numeric(1)}]\cr Strength of species-habitat association}

\item{complex}{[\code{Logical(1)}]\cr If TRUE, a clusters based at a regular grid is created (complex pattern)}
}
\value{
ppp object of the spatstat package with simulated species
}
\description{
Algorithm to create simulation pattern with six species with different species-habitat associations.
Species 1: No associations (Poisson process)
Species 2: Positive associations habitat 2 (Poisson process)
Species 3: Positive associations habitat 3 (Thomas process)
Species 4: Negative associations habitat 3 (Poisson process)
Species 5: Negative associations habitat 5 (Thomas process)
Species 6: No associations (Thomas process)
}
