% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_simulation_species.R
\name{Create.Simulation.Species}
\alias{Create.Simulation.Species}
\title{Create simulation species}
\usage{
Create.Simulation.Species(raster, type, process, habitat, number_points = 100,
  alpha = 0.3, species_code = 0, verbose = T)
}
\arguments{
\item{raster}{[\code{raster(1)}]\cr Raster object of the raster package with habitats}

\item{type}{[\code{string(1)}]\cr 'positive' or 'negative' associations}

\item{process}{[\code{string(1)}]\cr Process type to chose. Either 'Poisson' or 'Thomas'}

\item{habitat}{[\code{numeric(1)}]\cr Habitat to which species is associated}

\item{number_points}{[\code{numeric(1)}]\cr Number of points for each species (alpha=0)}

\item{alpha}{[\code{numeric(1)}]\cr Strength of species-habitat association}

\item{species_code}{[\code{numeric{1}}]\cr Species code to number species}

\item{verbose}{[\code{logical(1)}]\cr Print advanced error message}
}
\value{
ppp object of the spatstat package with simulated species
}
\description{
Algorithm to create simulation species with choosen characteristics
}
