% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neighbourhood.R
\name{create_neighbourhood}
\alias{create_neighbourhood}
\title{create_neighbourhood}
\usage{
create_neighbourhood(cells, matrix, directions = 4)
}
\arguments{
\item{cells}{matrix with cell ids of focal cells.}

\item{matrix}{matrix in which cells are located.}

\item{directions}{Cells neighbour rule: 4 (rook's case), 8 (queen's case).}
}
\value{
matrix
}
\description{
Create neighbourhood
}
\details{
Get cell ids of all neighbouring cells. The neighbourhoood rule can be specified
and is either rook's case (4 neighbours) or queen's case (8 neighbours).
}
\examples{
mat <- matrix(1, nrow= 10, ncol = 10)
cell_id <- rbind(cbind(3,5), cbind(7,1))
create_neighbourhood(cell_id, mat)

}
\seealso{
\code{\link{randomize_raster}}
}
