% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detection_habitat_associations.R
\name{Detection.Habitat.Association}
\alias{Detection.Habitat.Association}
\title{Detection of habitat association}
\usage{
Detection.Habitat.Association(input)
}
\arguments{
\item{input}{[\code{list(1)}]\cr List created with Results.Habitat.Associations}
}
\value{
Tibble with correct and false detections
}
\description{
Correct and false detections of the results of habitat associations
}
