% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_transf.R
\name{to_dummy}
\alias{to_dummy}
\alias{dummy_transf}
\title{dummy_transf}
\usage{
to_dummy(f)
}
\arguments{
\item{f}{Result of the calc_moments_full function which represents
product-moment contribution of a point at coordinates x, y with marks,
for the whole point pattern.}
}
\value{
matrix
}
\description{
Tranfsorm to dummy variables
}
\details{
Function for the transformation of variables to dummy variables and back
}
\keyword{internal}
