% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gest_multi.R
\name{estimate_gest_multi}
\alias{estimate_gest_multi}
\title{Multivariate G-function}
\usage{
estimate_gest_multi(pattern)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr ppp object of the spatstat packages containing multivariate point pattern}
}
\value{
Data frame containing r and the multivaria G function for all species separately and the overall mean
}
\description{
The multivariate nearest neighbour distribution function G(r) to all points belonging to a different species than the observed point
}
