% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pcf_fast.R
\name{estimate_pcf_fast}
\alias{estimate_pcf_fast}
\title{pcf-function}
\usage{
estimate_pcf_fast(pattern, ...)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr ppp object of the spatstat packages}

\item{...}{options passed to for Kest() and pcf.fv()}
}
\value{
fv object of the spatstat package
}
\description{
Computational fast method to calculate the pair correlation function using Ripleys K-Function
}
