% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pcf_fast_new.R
\name{estimate_pcf_fast_new}
\alias{estimate_pcf_fast_new}
\title{estimate_pcf_fast_new}
\usage{
estimate_pcf_fast_new(pattern, r = NULL, ...)
}
\arguments{
\item{pattern}{Point pattern.}

\item{r}{A vector specifying the maximum radius of the sample circles and
the interval length between successive sample circles radii.}

\item{...}{Arguments passed down to \code{kfun}.}
}
\value{
fv.object
}
\description{
Fast estimation of the pair correlation function
}
\details{
The functions estimates the pair correlation function using the \code{ads}
package.
}
\examples{
pcf_species_a <- estimate_pcf_fast_new(species_a)

}
\references{
Ripley, B.D. (1977) Modelling spatial patterns (with discussion). Journal of
the Royal Statistical Society, Series B, 39, 172-212.

Stoyan, D, Kendall, W.S. and Mecke, J. (1995) Stochastic geometry and its
applications. 2nd edition. Springer Verlag.

Stoyan, D. and Stoyan, H. (1994) Fractals, random shapes and point fields:
methods of geometrical statistics. John Wiley and Sons.
}
\seealso{
\code{\link{kfun}}
}
