% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pcf_multi.R
\name{estimate_pcf_multi}
\alias{estimate_pcf_multi}
\title{Multivariate pcf-function}
\usage{
estimate_pcf_multi(pattern)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr ppp object of the spatstat packages containing multivariate point pattern}
}
\value{
Data frame containing r and the multivaria pcf function for all species separately and the overall mean
}
\description{
The multivariate pair-correlation function pcf(r) to all points belonging to a different species than the observed point
}
