% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_points.R
\name{Extract.Points}
\alias{Extract.Points}
\title{Extract points within habitat}
\usage{
Extract.Points(raster, points, method)
}
\arguments{
\item{raster}{[\code{raster(1)}]\cr Raster object of the raster package.}

\item{points}{[\code{spatial points(1)}]\cr Spatial points of the sp package}

\item{method}{[\code{string(1)}]\cr Method which was used to produce results}
}
\value{
Table with counts in each habitat
}
\description{
Internal help function extracts the number of counts within each habitat
}
