% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_points.R
\name{extract_points}
\alias{extract_points}
\title{extract_points}
\usage{
extract_points(raster, pattern)
}
\arguments{
\item{raster}{RasterLayer}

\item{pattern}{Point pattern}
}
\value{
data.frame
}
\description{
Extract points
}
\details{
Extract number of pattern within each habitat
}
\examples{
\dontrun{
pattern_random <- spatstat::runifpoint(n = 50)
pcf_pattern_random <- estimate_pcf_fast(pattern_random)
}

}
