% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_point_process.R
\name{fit_point_process}
\alias{fit_point_process}
\title{Gamma test}
\usage{
fit_point_process(input, process = "poisson", number_pattern = 199)
}
\arguments{
\item{input}{[\code{ppp(1)}] ppp object of the spatstat package}

\item{process}{[\code{string(1)}] Method used for the reconstruction. Either "Poisson", "Cluster" or "Softcore"}

\item{number_pattern}{[\code{numeric(1)}] Number of reconstructed patterns}
}
\value{
List containing reconstructed patterns and observed pattern
}
\description{
Gamma test
}
