% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_pairs.R
\name{get_dist_pairs}
\alias{get_dist_pairs}
\title{get_dist_pairs}
\usage{
get_dist_pairs(X, rmax)
}
\arguments{
\item{X}{ppp object}

\item{rmax}{Numeric with maximum distance}
}
\value{
matrix
}
\description{
Distance between points
}
\details{
Returns matrix with point pairs and distances between them.
}
\seealso{
\code{\link{pcf.ppp}}
}
\keyword{internal}
