% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Habitat_classification.R
\name{Habitat.Classification}
\alias{Habitat.Classification}
\title{Habitat classification}
\usage{
Habitat.Classification(raster, classes = 5, style = "fisher")
}
\arguments{
\item{raster}{[\code{raster(1)}]\cr raster object of the raster package with continiuous habitats}

\item{classes}{[\code{numeric(1)}]\cr Number of classes}

\item{style}{[\code{string(1)}]\cr Style of breaks (see classInt::classInvervals() for more details)}
}
\value{
Raster object of the raster package with classified habitats
}
\description{
Classify continuous habitats values into discrete habitat types
}
