% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape}
\alias{landscape}
\title{Example landscape (random cluster neutral landscape model).}
\format{
A RasterLayer object.
}
\source{
Simulated neutral landscape model with R. https://github.com/ropensci/NLMR/
}
\usage{
landscape
}
\description{
An example map to show landscapetools functionality
generated with the \code{NLMR::nlm_fbm()} algorithm.
}
\keyword{datasets}
