% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_energy.R
\name{Mean.Energy}
\alias{Mean.Energy}
\title{Mean energy of the reconstruction}
\usage{
Mean.Energy(pattern, only_spatial = T)
}
\arguments{
\item{pattern}{[\code{list(1)}] List with reconstructed and observed point pattern created with SHAR::Pattern.Reconstruction()}

\item{only_spatial}{[\code{logical(1)}] TRUE for univariate point pattern}
}
\value{
Numeric with mean energy of the reconstruction
}
\description{
Function to plot all reconstructed patterns created of the pattern list with reconstructed patterns
}
