% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_complex.R
\name{Pattern.Complex}
\alias{Pattern.Complex}
\title{Help function}
\usage{
Pattern.Complex(owin)
}
\arguments{
\item{owin}{[\code{owin(1)}] Window of the spatstat package}
}
\value{
ppp object of the spatstat package
}
\description{
Internal help function to create complex pattern
}
