% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcf_fast.R
\name{Pcf.Fast}
\alias{Pcf.Fast}
\title{pcf-function}
\usage{
Pcf.Fast(pattern, correction = "good", method = "c", spar = 0.35)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr ppp object of the spatstat packages}

\item{correction}{[\code{string(1)}]\cr Edge correction (see ?spatstat::Kest() for more details)}

\item{method}{[\code{character(1)}]\cr See spatstat::pcf.fv() for mor details}

\item{spar}{[\code{numeric(1)}]\cr Smooting of spline function (see ?stats::smooth.spline() for more details)}
}
\value{
fv object of the spatstat package
}
\description{
Computational fast method to calculate the pair correlation function using Ripleys K-Function
}
