% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcf_multi.R
\name{Pcf.Multi}
\alias{Pcf.Multi}
\title{Multivariate pcf-function}
\usage{
Pcf.Multi(pattern, r_max, r_length = 515)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr ppp object of the spatstat packages containing multivariate point pattern}

\item{r.max}{[\code{numeric(1)}] Maximum distance r of multivariate G function}

\item{r.length}{[\code{numeric(1)}] Number of evaluated distances r}
}
\value{
Data frame containing r and the multivaria pcf function for all species separately and the overall mean
}
\description{
The multivariate pair-correlation function pcf(r) to all points belonging to a different species than the observed point
}
