% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_reconstructed_pattern.R
\name{Plot.Randomized.Pattern}
\alias{Plot.Randomized.Pattern}
\title{Plot list of randomized patterns}
\usage{
Plot.Randomized.Pattern(pattern, only_spatial = T, title = NULL)
}
\arguments{
\item{pattern}{[\code{list(1)}] List with reconstructed and observed point pattern created with SHAR::Pattern.Reconstruction()}

\item{only_spatial}{[\code{logical(1)}] TRUE for univariate point pattern}

\item{title}{[\code{string(1)}] Title of plot}
}
\value{
ggplot object of the ggplot2 package
}
\description{
Function to plot all randomized patterns created with either pattern reconstruction of the Gamma test
}
