% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rd_mar.R
\name{plot.rd_mar}
\alias{plot.rd_mar}
\title{plot.rd_mar}
\usage{
\method{plot}{rd_mar}(
  x,
  what = "sf",
  probs = c(0.025, 0.975),
  comp_fast = 1000,
  ask = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{rd_mar object with randomized patterns.}

\item{what}{Character specifying to plot summary functions of point patterns
(\code{what = "sf"}) or actual patterns (\code{what = "pp"}).}

\item{probs}{Vector with quantiles of randomized data used for envelope construction.}

\item{comp_fast}{Integer with threshold at which summary functions are estimated
in a computational fast way.}

\item{ask}{Logical if the user is asked to press <RETURN> before second summary function
is plotted (only used if \code{what = "sf"}).}

\item{verbose}{Logical if progress report is printed.}

\item{...}{Not used.}
}
\value{
void
}
\description{
Plot method for rd_pat object
}
\details{
The function plots the pair correlation function and the nearest neighbour function of
the observed pattern and the reconstructed patterns (as "simulation envelopes").
For large patterns \code{comp_fast = TRUE} decreases the computational demand because no edge
correction is used and the pair correlation function is estimated based on Ripley's
K-function. For more information see \code{\link{estimate_pcf_fast}}. It is also
possible to plot 3 randomized patterns and the observed pattern using \code{what = "pp"}.
}
\examples{
\dontrun{
pattern_recon <- reconstruct_pattern(species_a, n_random = 1, max_runs = 1000,
simplify = TRUE, return_input = FALSE)
marks_sub <- spatstat.geom::subset.ppp(species_a, select = dbh)
marks_recon <- reconstruct_pattern_marks(pattern_recon, marks_sub,
n_random = 19, max_runs = 1000)
plot(marks_recon)
}


}
\seealso{
\code{\link{reconstruct_pattern}} \cr
\code{\link{fit_point_process}}
}
