% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rd_multi.R
\name{plot.rd_multi}
\alias{plot.rd_multi}
\title{plot_sum_stat}
\usage{
\method{plot}{rd_multi}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{rd_multi Object created with \code{reconstruct_pattern_multi}.
multiple marks.}

\item{verbose}{Logical if progress should be printed.}

\item{...}{Currently not used}
}
\value{
void
}
\description{
plot_sum_stat
}
\details{
Calculates and visualises various summary statistics for the results of
multi-marks point pattern reconstruction.
}
\seealso{
\code{\link{reconstruct_pattern_multi}}
}
