% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_reconstructed_pattern.R
\name{Plot.Reconstructed.Pattern}
\alias{Plot.Reconstructed.Pattern}
\title{Plot list of reconstructed patterns}
\usage{
Plot.Reconstructed.Pattern(pattern, only_spatial = F)
}
\arguments{
\item{pattern}{[\code{list(1)}] List with reconstructed and observed point pattern created with SHAR::Pattern.Reconstruction()}

\item{only_spatial}{[\code{logical(1)}] TRUE for univariate point pattern}
}
\value{
ggplot object of the ggplot2 package
}
\description{
Function to plot all reconstructed patterns created of the pattern list with reconstructed patterns
}
