% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_randomized_pattern.R
\name{plot_randomized_pattern}
\alias{plot_randomized_pattern}
\title{Plot list of randomized patterns}
\usage{
plot_randomized_pattern(pattern, only_spatial = TRUE, title = NULL)
}
\arguments{
\item{pattern}{[\code{list(1)}] List with reconstructed and observed point pattern created with SHAR::Pattern.Reconstruction()}

\item{only_spatial}{[\code{logical(1)}] TRUE for univariate point pattern}

\item{title}{[\code{string(1)}] Title of plot}
}
\value{
ggplot object of the ggplot2 package
}
\description{
Function to plot all randomized patterns created with either pattern reconstruction of the Gamma test
}
