% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_randomized_raster.R
\name{plot_randomized_raster}
\alias{plot_randomized_raster}
\title{plot_randomized_raster}
\usage{
plot_randomized_raster(raster, n = NULL, col, verbose = TRUE, nrow, ncol)
}
\arguments{
\item{raster}{rd_ras object with randomized raster.}

\item{n}{Integer with number or vector of ids of randomized raster to plot.
See Details section for more information.}

\item{col}{Vector with color palette used for plotting.}

\item{verbose}{Logical if messages are printed.}

\item{nrow, ncol}{Integer with number of rows and columns of plot grid.}
}
\value{
void
}
\description{
Plot randomized raster
}
\details{
Function to plot randomized raster. If \code{n} is a single number, \code{n} randomized
raster will be sampled to plot. If \code{n} is a vector, the corresponding raster
will be plotted. \code{col, nrow, ncol} are passed to \code{plot}.
}
\examples{
\dontrun{
landscape_classes <- classify_habitats(raster = landscape, classes = 5)
landscape_random <- randomize_raster(raster = landscape_classes, n_random = 19)

plot_randomized_raster(landscape_random)

palette <- viridis::viridis(n = 5)
plot_randomized_raster(landscape_random, n = 5, col = palette, nrow = 3, ncol = 2)
}

}
\seealso{
\code{\link{randomize_raster}} \cr
\code{\link{translate_raster}}
}
