% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sum_stat.R
\name{plot_sum_stat}
\alias{plot_sum_stat}
\title{plot_sum_stat}
\usage{
plot_sum_stat(reconstruction)
}
\arguments{
\item{reconstruction}{Result list of the dot pattern reconstruction with
multiple marks.}
}
\value{
ggplot object
}
\description{
plot_sum_stat
}
\details{
Calculates and visualises various summary statistics for the results of
multi-marks point pattern reconstruction.
}
