% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rd_pat.R
\name{print.rd_pat}
\alias{print.rd_pat}
\title{print.rd_pat}
\usage{
\method{print}{rd_pat}(x, digits = 4, ...)
}
\arguments{
\item{x}{rd_pat object with randomized patterns.}

\item{digits}{Integer with number of decimal places (round).}

\item{...}{Arguments passed to \code{cat}.}
}
\value{
void
}
\description{
Print method for rd_pat object
}
\details{
Printing method for random patterns created with \code{reconstruct_pattern_*}.
}
\examples{
pattern_random <- fit_point_process(species_a, n_random = 199)
print(pattern_random)

\dontrun{
pattern_recon <- reconstruct_pattern(species_b, n_random = 19, max_runs = 1000,
method = "hetero")
print(pattern_recon)
}

}
\seealso{
\code{\link{reconstruct_pattern}} \cr
\code{\link{fit_point_process}}
}
