% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rd_ras.R
\name{print.rd_ras}
\alias{print.rd_ras}
\title{print.rd_ras}
\usage{
\method{print}{rd_ras}(x, ...)
}
\arguments{
\item{x}{rd_ras object with randomized raster.}

\item{...}{Arguments passed to \code{cat}.}
}
\value{
void
}
\description{
Print method for rd_ras object
}
\details{
Printing method for random patterns created with \code{\link{randomize_raster}} or
\code{\link{translate_raster}}.
}
\examples{
\dontrun{
landscape_classified <- classify_habitats(terra::rast(landscape), n = 5, style = "fisher")
landscape_random <- randomize_raster(landscape_classified, n_random = 19)

print(landscape_random)
}

}
\seealso{
\code{\link{randomize_raster}} \cr
\code{\link{translate_raster}}
}
