% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_walk}
\alias{random_walk}
\title{Random walk}
\format{rd_ras object.}
\usage{
random_walk
}
\description{
Randomization of the landscape data using the habitat randomization algorithm.
}
\keyword{datasets}
