% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomization_algorithm.R
\name{Randomization.Algorithm}
\alias{Randomization.Algorithm}
\title{Internal help function}
\usage{
Randomization.Algorithm(raster, number_neighbours = 8)
}
\arguments{
\item{raster}{[\code{raster(1)}] Raster of the raster object with discrete habitats}

\item{number_neighbours}{[\code{numeric(1)}] Number of random neigbhours. See ?raster::adjacent() for more details}
}
\value{
Raster of the raster package with randomized habitats
}
\description{
Randomization algorithm assigning habitat to random neighbours
}
