% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization_algorithm.R
\name{randomization_algorithm}
\alias{randomization_algorithm}
\title{randomization_algorithm}
\usage{
randomization_algorithm(raster, n_random = 19, direction = 8,
  verbose = FALSE)
}
\arguments{
\item{raster}{RasterLayer.}

\item{n_random}{Number of randomized RasterLayers.}

\item{direction}{Number of directions in which cells should be connected:
4 (rook's case), 8 (queen's case), 16 (knight and one-cell queen moves),
or 'bishop' to connect cells with one-cell diagonal moves. Or a neigborhood matrix.}

\item{verbose}{Print progress report.}
}
\value{
list
}
\description{
Randomization algorithm
}
\details{
The function randomizes a habitat map (as RasterLayer) as proposed by Harms et al. (2001)
as “randomized-habitats procedure”. The algorithm starts with an empty habitat map
starts to assign random neighbouring cells (specified by `directions`) to each
habitat (in increasing order of abundance in observed map). We modified the
procedure slightly by increasing a probability to jump to a non-neighbouring
cell as the current patch becomes larger.
}
\examples{
\dontrun{
landscape <- NLMR::nlm_fbm(ncol = 50, nrow = 50, user_seed = 1)
landscape_classified <- SHAR::classify_habitats(landscape, classes = 5)
landscape_random <- randomization_algorithm(landscape_classified, n_random = 9)
}

}
\references{
Harms, K. E., Condit, R., Hubbell, S. P., & Foster, R. B. (2001). Habitat associations
of trees and shrubs in a 50-ha neotropical forest plot. Journal of Ecology, 89(6), 947–959.
}
\seealso{
\code{\link{translate_raster}} \cr
\code{\link{adjacent}}
}
