% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_sample}
\alias{rcpp_sample}
\title{rcpp_sample}
\usage{
rcpp_sample(x, n, replace = FALSE)
}
\arguments{
\item{x}{Vector of elements to sample from.}

\item{n}{Integer with size of the sample.}

\item{replace}{Logical if sample with replacement.}
}
\value{
vector
}
\description{
Rcpp sample function
}
\details{
\code{Rcpp} implementation of the \code{sample} function.
}
\seealso{
\code{\link{sample}}
}
\keyword{internal}
