% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labeling_reconstruction.R
\name{reconstruct_labeling}
\alias{reconstruct_labeling}
\title{Pattern reconstruction}
\usage{
reconstruct_labeling(pattern, max_runs = 10000, e_threshold = 0.01,
  fitting = FALSE)
}
\arguments{
\item{pattern}{[\code{ppp(1)}] Multivariate ppp object of the spatstat package}

\item{max_runs}{[\code{numeric(1)}] number of maximum iterations}

\item{e_threshold}{[\code{numeric(1)}] Threshold for energy to reach during reconstruction}

\item{fitting}{[\code{logical(1)}] If TRUE, a clustered pattern is fitted to the original pattern as starting point}
}
\value{
ppp object of the spatstat package with reconstructed pattern
}
\description{
Multivariate pattern reconstruction using a random marking approach
}
