% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_marks.R
\name{reconstruct_marks}
\alias{reconstruct_marks}
\title{reconstruct_marks}
\usage{
reconstruct_marks(pattern, marked_pattern, mark, n_random = 19,
  e_threshold = 0.01, max_runs = 10000, return_input = TRUE,
  verbose = FALSE, plot = FALSE)
}
\arguments{
\item{pattern}{ppp.}

\item{n_random}{Number of randomizations.}

\item{e_threshold}{Minimum energy to stop reconstruction.}

\item{max_runs}{Maximum number of iterations of e_threshold is not reached.}

\item{return_input}{The original input data is returned as last list entry}

\item{verbose}{Print progress report.}

\item{plot}{Plot mark function during optimization}
}
\value{
list
}
\description{
Pattern reconstruction of marks
}
\details{

}
\examples{
\dontrun{
species_a
species_b

pattern_recon <- reconstruct_marks(species_b, n_random = 39, max_runs = 1000)
}

}
\references{
Tscheschel, A., & Stoyan, D. (2006). Statistical reconstruction of random point
patterns. Computational Statistics and Data Analysis, 51(2), 859–871.

Wiegand, T., & Moloney, K. A. (2014). Handbook of spatial point-pattern analysis
in ecology. Boca Raton: Chapman and Hall/CRC Press.
}
\seealso{
\code{\link{reconstruct_pattern}} \cr
}
