% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_pattern_cluster.R
\name{reconstruct_pattern_cluster}
\alias{reconstruct_pattern_cluster}
\title{reconstruct_pattern_cluster}
\usage{
reconstruct_pattern_cluster(
  pattern,
  n_random = 1,
  e_threshold = 0.01,
  max_runs = 1000,
  no_change = Inf,
  annealing = 0.01,
  comp_fast = 1000,
  weights = c(0.5, 0.5),
  r_length = 250,
  return_input = TRUE,
  simplify = FALSE,
  verbose = TRUE,
  plot = FALSE
)
}
\arguments{
\item{pattern}{ppp object with pattern.}

\item{n_random}{Integer with number of randomizations.}

\item{e_threshold}{Double with minimum energy to stop reconstruction.}

\item{max_runs}{Integer with maximum number of iterations if \code{e_threshold}
is not reached.}

\item{no_change}{Integer with number of iterations at which the reconstruction will
stop if the energy does not decrease.}

\item{annealing}{Double with probability to keep relocated point even if energy
did not decrease.}

\item{comp_fast}{Integer with threshold at which summary functions are estimated
in a computational fast way.}

\item{weights}{Vector with weights used to calculate energy.
The first number refers to Gest(r), the second number to pcf(r).}

\item{r_length}{Integer with number of intervals from \code{r = 0} to \code{r = rmax} for which
the summary functions are evaluated.}

\item{return_input}{Logical if the original input data is returned.}

\item{simplify}{Logical if only pattern will be returned if \code{n_random = 1}
and \code{return_input = FALSE}.}

\item{verbose}{Logical if progress report is printed.}

\item{plot}{Logical if pcf(r) function is plotted and updated during optimization.}
}
\value{
rd_pat
}
\description{
Pattern reconstruction for clustered patterns
}
\details{
The functions randomizes the observed pattern by using pattern reconstruction
as described in Tscheschel & Stoyan (2006) and Wiegand & Moloney (2014). The
algorithm starts with a random but clustered pattern, shifts a point to a new location and
keeps the change only, if the deviation between the observed and the reconstructed
pattern decreases. The pair correlation function and the nearest neighbour
distance function are used to describe the patterns.

For large patterns (\code{n > comp_fast}) the pair correlation function can be estimated
from Ripley's K-function without edge correction. This decreases the computational
time. For more information see \code{\link{estimate_pcf_fast}}.

The reconstruction can be stopped automatically if for n steps the energy does not
decrease. The number of steps can be controlled by \code{no_change} and is set to
\code{no_change = Inf} as default to never stop automatically.

The weights must be 0 < sum(weights) <= 1. To weight both summary functions identical,
use \code{weights = c(0.5, 0.5)}.

\code{spatstat} sets \code{r_length} to 513 by default. However, a lower value decreases
the computational time, while increasing the "bumpiness" of the summary function.
}
\examples{
\dontrun{
pattern_recon <- reconstruct_pattern_cluster(species_b, n_random = 19, max_runs = 1000)
}

}
\references{
Tscheschel, A., & Stoyan, D. (2006). Statistical reconstruction of random point
patterns. Computational Statistics and Data Analysis, 51(2), 859-871.

Wiegand, T., & Moloney, K. A. (2014). Handbook of spatial point-pattern analysis
in ecology. Boca Raton: Chapman and Hall/CRC Press.
}
\seealso{
\code{\link{calculate_energy}} \cr
\code{\link{plot_randomized_pattern}}
\code{\link{reconstruct_pattern_homo}} \cr
\code{\link{reconstruct_pattern_hetero}} \cr
\code{\link{reconstruct_pattern_marks}}
}
