% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_pattern_multi.R
\name{reconstruct_pattern_multi}
\alias{reconstruct_pattern_multi}
\title{reconstruct_pattern_multi}
\usage{
reconstruct_pattern_multi(
  marked_pattern,
  xr = marked_pattern$window$xrange,
  yr = marked_pattern$window$yrange,
  n_repetitions = 1,
  max_steps = 10000,
  no_changes = 5,
  rcount = 250,
  rmax = 25,
  issue = 1000,
  divisor = "r",
  kernel_arg = "epanechnikov",
  timing = FALSE,
  energy_evaluation = FALSE,
  show_graphic = FALSE,
  Lp = 1,
  bw = if (divisor \%in\% c("r", "d")) 0.5 else 5,
  sd = "step",
  steps_tol = 1000,
  tol = 1e-04,
  w_markcorr = c(d_d = 1, all = 1, d_all = 1, all_all = 1, d_d0 = 1, all0 = 1, d_all0 =
    1, all_all0 = 1),
  prob_of_actions = c(move_coordinate = 0.4, switch_coords = 0.1, exchange_mark_one =
    0.1, exchange_mark_two = 0.1, pick_mark_one = 0.2, pick_mark_two = 0.1),
  k = 1,
  w_statistics = c()
)
}
\arguments{
\item{marked_pattern}{ppp  object with marked pattern. See Details section
for more information.}

\item{xr, yr}{Add description}

\item{n_repetitions}{Integer representing the number of simulations to be
performed.}

\item{max_steps}{Maximum number simulation steps.}

\item{no_changes}{Integer representing the number of iterations
(per 1000 simulation steps)
after which the reconstruction is terminated if the energy does not decrease.}

\item{rcount}{Integer representing the number of intervals for which the
summary statistics are evaluated.}

\item{rmax}{Maximum distance [m] at which the summary statistics are
evaluated.}

\item{issue}{Integer that determines after how many simulations steps an
output occurs.}

\item{divisor}{Divisor in the smoothing kernel, d or r.}

\item{kernel_arg}{The kernel used to calculate the energy, possible kernels
can be: Gaussian, Epanechnikov, Rectangular, Cumulative.}

\item{timing}{Logical value: The computation timeis measured if this is TRUE}

\item{energy_evaluation}{Logical value: If this is TRUE, the procedure stores
the energy shares of the total energy per simulation step.}

\item{show_graphic}{Logical value: If this is TRUE, the procedure records the
point pattern during optimisation and updated}

\item{Lp}{Distance measure for the calculation of the energy function
(Lp distance, 1 ≤ p < ∞).}

\item{bw}{Bandwidth [m] with which the kernels are scaled, so that this is
the standard deviation of the smoothing kernel.}

\item{sd}{This is the standard deviation [m] used in the move_coordinate
action.}

\item{steps_tol}{After the value steps_tol it is checked whether the energy
change is smaller than tol.}

\item{tol}{Tolerance: stops the procedure of energy more than 1 − tol
times no_changes.}

\item{w_markcorr}{Vector of possible weightings of individual mcf's.
\code{c(1, 1, 1, 1, 1, 1, 1, 1)}}

\item{prob_of_actions}{Vector of probabilities for the actions performed.
\code{c(move_coordinate = 0.4, switch_coords = 0.1, exchange_mark_one = 0.1,
exchange_mark_two = 0.1, pick_mark_one = 0.2, pick_mark_two = 0.1)}}

\item{k}{Vector of values k; used only if Dk is included in w_statistics
elow.}

\item{w_statistics}{vector of named weights for optional spatial statistics
from the ‘spatstat’package to be included in the energy calculation.This may
include Dk, K, Hs, pcf.}
}
\value{
rd_mar
}
\description{
Pattern reconstruction of a pattern marked by multiple traits.
}
\details{
A novel approach carries out a pattern reconstruction of marked dot patterns
as described by Tscheschel and Stoyan (2006) and Wiegand and Moloney (2014).

One particular feature is the simultaneous consideration of both marks,
accounting for their correlation during reconstruction.

The marked point pattern (PPP object) must is currently structured as follows:
X-coordinate, Y-coordinate, metric mark (e.g. diameter at breast height),
and nominal mark (e.g. tree species).It is calculated in the unit metre [m].

A combination of the mark correlation function and pair correlation function
is used for pattern description. Additional summary statistics may be
considered.Two randomly selected marks are chosen in each iteration, and one
of various actions is performed. Changes will only be retained if the
difference between the observed and reconstructed pattern decreases
(minimizing energy).

This method is currently only suitable for homogeneous point patterns.

A comprehensive description of the method can be
found in Wudel et al. (2023). Future developments will involve considering
more than two markers, expanding and reducing the observation window under
using an Edge Correction. Additionally, data input will be automatically
detected.
}
\examples{
\dontrun{

# Random example data set
  xr <- 500
  yr <- 1000
  N <- 400
  y <- runif(N, min = 0, max = yr)
  x <- runif(N, min = 0, max = xr)
  species <- as.factor(c("A","B"))
  species <- sample(species, N, replace = TRUE)
  diameter <- runif(N, 1, 40)
  plot(x,y, col = species, pch = 19, cex = diameter*0.02)
  random <- data.frame(x, y, diameter/ 0.1,species)
  colnames(random) <- c("x", "y", "dbh [mm]", "Tree species")

# Conversion to a ppp object and conversion of the metric mark to metres.
  w <- owin(c(0, 500), c(0, 1000))
  marked_pattern <- as.ppp(data.frame(random), W = w)
  marked_pattern$marks$dbh..mm.<-marked_pattern$marks$dbh..mm.*0.001

# Reconstruction function
reconstruction <- reconstruct_pattern_multi(
  marked_pattern, n_repetitions = 1, max_steps = 100000, no_changes = 5,
  rcount = 250, rmax = 25, issue = 1000, divisor = "r",
  kernel_arg = "epanechnikov", timing = TRUE, energy_evaluation = TRUE,
  show_graphic = TRUE, Lp = 1, bw = 0.5, sd = "step", steps_tol = 1000,
  tol = 1e-4, w_markcorr = c(d_d=1, all=1, d_all=1, all_all=1, d_d0=1,
  all0=1, d_all0=1, all_all0=1), prob_of_actions = c(move_coordinate = 0.4,
  switch_coords = 0.1, exchange_mark_one = 0.1, exchange_mark_two = 0.1,
  pick_mark_one = 0.2, pick_mark_two = 0.1),
  k = 1, w_statistics = c()
)
}

}
\references{
Kirkpatrick, S., Gelatt, C.D.Jr., Vecchi, M.P., 1983. Optimization by simulated
annealing. Science 220, 671–680. <https://doi.org/10.1126/science.220.4598.671>

Tscheschel, A., Stoyan, D., 2006. Statistical reconstruction of random point
patterns. Computational Statistics and Data Analysis 51, 859–871.
<https://doi.org/10.1016/j.csda.2005.09.007>

Wiegand, T., Moloney, K.A., 2014. Handbook of spatial point-pattern analysis in
ecology. Chapman and Hall/CRC Press, Boca Raton. ISBN 978-1-4200-8254-8

Wudel, C., Schlicht, R., & Berger, U. (2023). Multi-trait point pattern
reconstruction of plant ecosystems. Methods in Ecology and Evolution, 14, 2668–2679.
https://doi.org/10.1111/2041-210X.14206
}
\seealso{
\code{\link{fit_point_process}} \cr
\code{\link{reconstruct_pattern}} \cr
\code{\link{reconstruct_pattern_marks}}
}
