% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_reconstruction.R
\name{reconstruct_spatial}
\alias{reconstruct_spatial}
\title{Spatial reconstruction}
\usage{
reconstruct_spatial(pattern, max_runs = 10000, e_threshold = 0.01,
  fitting = FALSE)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr Multivariate ppp object of the spatstat package}

\item{max_runs}{[\code{numeric(1)}]\cr number of maximum iterations}

\item{e_threshold}{[\code{numeric(1)}]\cr Threshold for energy to reach during reconstruction}

\item{fitting}{[\code{logical(1)}]\cr If TRUE, a clustered pattern is fitted to the original pattern as starting point}
}
\value{
ppp object of the spatstat package with reconstructed pattern
}
\description{
Univariate pattern reconstruction of only the spatial structure
}
