% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reconstruction}
\alias{reconstruction}
\title{Reconstruction}
\format{rd_pat object.}
\usage{
reconstruction
}
\description{
Randomized data for species b using pattern reconstruction.
}
\keyword{datasets}
