% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_habitat_association.R
\name{Results.Habitat.Association}
\alias{Results.Habitat.Association}
\title{Results habitat association}
\usage{
Results.Habitat.Association(pattern, raster, method, threshold = c(0.025,
  0.975), only_spatial = F)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr ppp object of the spatstat package. If method="random_pattern", list with simulated patterns}

\item{raster}{[\code{raster(1)}]\cr Raster object of the raster package. If method="random_raster, list with simulated habitats}

\item{method}{[\code{string(1)}]\cr Form of input data. Either "random_pattern " or "random_raster"}

\item{threshold}{[\code{numeric(2)}]\cr Threshold for significance. See ?quantile() for more details}

\item{only_spatial}{[\code{logical(1)}]\cr Boolean if pattern contains only one species}
}
\value{
List or data frame with results of species-habitat association
}
\description{
Results of the species-habitat association test using the simulated vs. the observed data
}
