% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_randomized.R
\name{sample_randomized}
\alias{sample_randomized}
\title{sample_randomized}
\usage{
sample_randomized(randomized, n = NULL, verbose = TRUE)
}
\arguments{
\item{randomized}{List with randomized raster or patterns.}

\item{n}{Integer with number or vector of ids of randomized pattern to plot.}

\item{verbose}{Logical if progress report is printed.}
}
\value{
list
}
\description{
Sample randomized list
}
\details{
Get list with \code{n} randomized raster or patterns. If \code{n} is a single number,
\code{n} randomized elements will be sampledt. If \code{n} is a vector, the
corresponding elements will be returned.
}
\examples{
\dontrun{
sample_randomized(randomized = reconstruction$randomized, n = c(5, 10, 15))
}

}
\keyword{internal}
