% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_species.R
\name{Select.Species}
\alias{Select.Species}
\title{Internal help function}
\usage{
Select.Species(pattern)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr ppp object of the spatstat packages containing multivariate point pattern}
}
\description{
Internal help function to select pattern with only species as mark
}
