% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_kernel.R
\name{select_kernel}
\alias{select_kernel}
\title{select_kernel}
\usage{
select_kernel(kernel_arg, bw, rmax, divisor)
}
\arguments{
\item{kernel_arg}{Parameter of the function
reconstruct_pattern_multi_trait_marks, specifies the kernel to be used to
calculate the energy, possible kernels can be: Gaussian, Epanechnikov,
Rectangular, Cumulative.}

\item{bw}{Bandwidth with which the kernels are scaled, so that this is the
standard deviation of the smoothing kernel.}

\item{rmax}{Maximum distance at which the summary statistics are evaluated.}

\item{divisor}{Divisor in the smoothing kernel, d or r.}
}
\value{
list
}
\description{
Kernel selection
}
\details{
Returns the function of the selected kernel, which is then used to
calculate the kernel.
}
\keyword{internal}
