% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_species.R
\name{select_species}
\alias{select_species}
\title{Internal help function}
\usage{
select_species(pattern)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr ppp object of the spatstat packages containing multivariate point pattern}
}
\description{
Internal help function to select pattern with only species as mark
}
