% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAR.R
\docType{package}
\name{SHAR-package}
\alias{SHAR}
\alias{SHAR-package}
\title{SHAR}
\description{
\emph{SHAR} provides ,,,
}
\author{
\strong{Maintainer}: Maximilian H.K. Hesselbarth \email{maximilian.hesselbarth@uni-goettingen.de}

}
