% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shar-package.R
\docType{package}
\name{shar}
\alias{shar-package}
\alias{shar}
\title{Package description}
\description{
Analyse species-habitat associations in R. Therefore, information about the
location of the species is needed and about the environmental conditions. To test
for significance habitat associations, one of the two components is randomized.
Methods are mainly based on Plotkin et al. (2000) <doi:10.1006/jtbi.2000.2158> and
Harms et al. (2001) <doi:10.1111/j.1365-2745.2001.00615.x>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-spatialecology.github.io/shar/}
  \item Report bugs at \url{https://github.com/r-spatialecology/shar/issues/}
}

}
\author{
\strong{Maintainer}: Maximilian H.K. Hesselbarth \email{mhk.hesselbarth@gmail.com} (\href{https://orcid.org/0000-0003-1125-9918}{ORCID})

Authors:
\itemize{
  \item Marco Sciaini \email{marco.sciaini@posteo.net} (\href{https://orcid.org/0000-0002-3042-5435}{ORCID})
  \item Chris Wudel \email{chris.wudel@tu-dresden.de} (\href{https://orcid.org/0000-0003-0446-4665}{ORCID})
}

Other contributors:
\itemize{
  \item Zeke Marshall \email{ee18zm@leeds.ac.uk} (\href{https://orcid.org/0000-0001-9260-7827}{ORCID}) [contributor]
  \item Thomas Etherington \email{teth001@aucklanduni.ac.nz} (\href{https://orcid.org/0000-0002-3187-075X}{ORCID}) [contributor]
  \item Janosch Heinermann \email{janosch.heinermann@tu-dresden.de} (\href{https://orcid.org/0000-0001-5080-8591}{ORCID}) [contributor]
}

}
\keyword{internal}
