% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_a}
\alias{species_a}
\title{Species a}
\format{
A spatstat ppp object.
}
\usage{
species_a
}
\description{
A species with negative associations to habitat 4 of \code{landscape}. Please be
aware that a negative association to one habitat will inevitable lead to positive
associations to other habitats (Yamada et al. 2006).
}
\references{
Yamada, T., Tomita, A., Itoh, A., Yamakura, T., Ohkubo, T., Kanzaki, M., Tan, S.,
Ashton, P.S., 2006. Habitat associations of Sterculiaceae trees in a Bornean rain
forest plot. Journal of Vegetation Science 17, 559–566.
}
\keyword{datasets}
