% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_b}
\alias{species_b}
\title{Species b}
\format{
A spatstat ppp object.
}
\usage{
species_b
}
\description{
A species with positive associations to habitat 5 of \code{landscape}. Please be
aware that a positive association to one habitat will inevitable lead to negative
associations to other habitats (Yamada et al. 2006)
}
\references{
Yamada, T., Tomita, A., Itoh, A., Yamakura, T., Ohkubo, T., Kanzaki, M., Tan, S.,
Ashton, P.S., 2006. Habitat associations of Sterculiaceae trees in a Bornean rain
forest plot. Journal of Vegetation Science 17, 559–566.
https://doi.org/10.1111/j.1654-1103.2006.tb02479.x
}
\keyword{datasets}
