% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Torus_translation_raster.R
\name{Torus.Translation}
\alias{Torus.Translation}
\title{Torus translation}
\usage{
Torus.Translation(raster, x_shift = 0, y_shift = 0)
}
\arguments{
\item{raster}{[\code{raster(1)}] Raster of the raster package}

\item{x_shift}{[\code{numeric(1)}] Number of pixels shifted in x direction}

\item{y_shift}{[\code{numeric(1)}] Number of pixels shifted in y direction}
}
\value{
Raster object of the raster package
}
\description{
Shifts the raster in x or y direction on a tours
}
