% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_raster.R
\name{translate_raster}
\alias{translate_raster}
\title{translate_raster}
\usage{
translate_raster(raster)
}
\arguments{
\item{raster}{RasterLayer or list of RasterLayers}
}
\value{
list
}
\description{
Torus translation
}
\details{
Torus translation
}
\examples{
\dontrun{
landscape <- NLMR::nlm_fbm(ncol = 50, nrow = 50, user_seed = 1)
landscape_classified <- SHAR::classify_habitats(landscape, classes = 5)
landscape_random <- translate_raster(landscape_classified)
}

}
\references{
Harms, K. E., Condit, R., Hubbell, S. P., & Foster, R. B. (2001). Habitat associations
of trees and shrubs in a 50-ha neotropical forest plot. Journal of Ecology, 89(6), 947–959.
}
\seealso{
\code{\link{randomization_algorithm}}
}
