% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_energy.R
\name{calculate_energy}
\alias{calculate_energy}
\title{Help function to calculate enegry}
\usage{
calculate_energy(pattern, gest_observed, pcf_observed, only_spatial)
}
\arguments{
\item{pattern}{[\code{list(1)}] List with reconstructed point patterns created with SHAR::Pattern.Reconstruction()}

\item{gest_observed}{[\code{fv(1)}] Neareast neighbour distribution functions o observed pattern}

\item{pcf_observed}{[\code{fv(1)}] Pair correlation function of observed pattern}

\item{only_spatial}{[\code{logical(1)}] TRUE for univariate point pattern}
}
\description{
Help function to calculate enegry
}
