% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_energy.R
\name{calculate_mean_energy}
\alias{calculate_mean_energy}
\title{calculate_mean_energy}
\usage{
calculate_mean_energy(pattern, comp_fast = FALSE)
}
\arguments{
\item{pattern}{List with reconstructed patterns}

\item{comp_fast}{Logical if summary functions should be estimated in an computational
fast way (but without edge correction)}
}
\value{
numeric
}
\description{
Calculate mean energy
}
\details{
Calculate the mean energy between the observed pattern and all reconstructed patterns.
}
\examples{
\dontrun{
pattern_random <- spatstat::runifpoint(n = 50)
pattern_recon <- SHAR::reconstruct_pattern(pattern_random, n_random = 9, max_runs = 1000)
calculate_mean_energy(pattern_recon)
}

}
\references{
Tscheschel, A., & Stoyan, D. (2006). Statistical reconstruction of random point
patterns. Computational Statistics and Data Analysis, 51(2), 859–871.

Wiegand, T., & Moloney, K. A. (2014). Handbook of spatial point-pattern analysis
in ecology. Boca Raton: Chapman and Hall/CRC Press.
}
\seealso{
\code{\link{plot_randomized_pattern}} \cr
\code{\link{reconstruct_pattern}}
}
