% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_statistics.R
\name{compute_statistics}
\alias{compute_statistics}
\title{compute_statistics}
\usage{
compute_statistics(x, y, k, xr, yr, w_statistics, bw, divisor, kernel_arg, r)
}
\arguments{
\item{x, y}{x and y coordinates of the points from the reference point pattern.}

\item{k}{Vector of values k; used only if Dk is included in w_statistics below.}

\item{xr, yr}{x and y extension of the observation window (start, end).}

\item{w_statistics}{vector of named weights for optional spatial statistics
from the \code{spatstat} package to be included in the energy calculation. This may
include Dk, K, Hs, pcf.}

\item{bw, divisor, kernel_arg, r}{Several parameters related to summary function.}
}
\value{
list
}
\description{
Compute summary statistics
}
\details{
Compute optional spatial statistics using the spatstat package.
}
\keyword{internal}
