% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_simulation_pattern.R
\name{Create.Simulation.Pattern}
\alias{Create.Simulation.Pattern}
\title{Create simulation pattern}
\usage{
Create.Simulation.Pattern(raster, number_points = 100, alpha = 0.3)
}
\arguments{
\item{raster}{[\code{raster(1)}]\cr Raster object of the raster package with habitats}

\item{number_points}{[\code{numeric(1)}]\cr Number of points for each species (alpha=0)}

\item{alpha}{[\code{numeric(1)}]\cr Strength of species-habitat association}
}
\value{
ppp object of the spatstat package with simulated species
}
\description{
Algorithm to create simulation pattern with six species with different species-habitat associations.
\cr Species 1: Positive associations (Poisson process)
\cr Species 2: Positive associations (Thomas process)
\cr Species 3: Negative associations (Poisson process)
\cr Species 4: Negative associations (Thomas process)
}
