% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pcf_fast.R
\name{estimate_pcf_fast}
\alias{estimate_pcf_fast}
\title{estimate_pcf_fast}
\usage{
estimate_pcf_fast(pattern, ...)
}
\arguments{
\item{pattern}{Point pattern}

\item{...}{Arguments passed to \code{\link{Kest}} and \code{\link{pcf.fv}}}
}
\value{
fv.object
}
\description{
Fast estimation of the pair correlation function
}
\details{
Classifies a RasterLayer with continious values into n discrete classes
}
\examples{
\dontrun{
pattern_random <- spatstat::runifpoint(n = 50)
pcf_pattern_random <- estimate_pcf_fast(pattern_random)
}

}
\seealso{
\code{\link{Kest}} \cr
\code{\link{pcf.fv}}
}
