% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pcf_fast.R
\name{estimate_pcf_fast}
\alias{estimate_pcf_fast}
\title{estimate_pcf_fast}
\usage{
estimate_pcf_fast(pattern, ...)
}
\arguments{
\item{pattern}{ppp object with point pattern.}

\item{...}{Arguments passed down to \code{link{Kest}} or \code{\link{pcf.fv}}.}
}
\value{
fv.object
}
\description{
Fast estimation of the pair correlation function
}
\details{
The functions estimates the pair correlation functions based on an estimation
of Ripley's K-function. This makes it computationally faster than estimating the
pair correlation function directly. It is a wrapper around \code{\link{Kest}} and
\code{\link{pcf.fv}}.
}
\examples{
pcf_species_b <- estimate_pcf_fast(species_a)

}
\references{
Chiu, S.N., Stoyan, D., Kendall, W.S., Mecke, J., 2013. Stochastic geometry and
its applications, 3rd ed, Wiley Series in Probability and Statistics.
John Wiley & Sons Inc, Chichester, UK. isbn:978-0-470-66481-0

Ripley, B.D., 1977. Modelling spatial patterns. Journal of the Royal Statistical
Society. Series B (Methodological) 39, 172–192.
<https://doi.org/10.1111/j.2517-6161.1977.tb01615.x>

Stoyan, D., Stoyan, H., 1994. Fractals, random shapes and point fields.
John Wiley & Sons, Chichester. isbn:978-0-471-93757-9
}
\seealso{
\code{\link{Kest}} \cr
\code{\link{pcf.fv}}
}
