% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_points.R
\name{extract_points}
\alias{extract_points}
\title{extract_points}
\usage{
extract_points(raster, pattern)
}
\arguments{
\item{raster}{RasterLayer with environmental data}

\item{pattern}{ppp object with point pattern.}
}
\value{
data.frame
}
\description{
Extract points
}
\details{
The function extracts the number of points within each discrete habitat.
}
\examples{
\dontrun{
landscape_classified <- classify_habitats(landscape, n = 5, style = "fisher")
extract_points(raster = landscape_classified, pattern = species_b)
}

}
\seealso{
\code{\link{results_habitat_association}}
}
\keyword{internal}
