% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_points.R
\name{extract_points}
\alias{extract_points}
\title{extract_points}
\usage{
extract_points(raster, pattern)
}
\arguments{
\item{raster}{RasterLayer.}

\item{pattern}{Point pattern.}
}
\value{
data.frame
}
\description{
Extract points
}
\details{
The function extracts the number of points within each habitat.
}
\examples{
\dontrun{
landscape <- NLMR::nlm_fbm(ncol = 50, nrow = 50, user_seed = 1)
landscape_classified <- SHAR::classify_habitats(landscape, classes = 5)
species_1 <- spatstat::runifpoint(n = 50, win = spatstat::owin(c(0, 50), c(0, 50)))
extract_points(raster = landscape_classified, pattern = species_1)
}

}
