% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_points.R
\name{extract_points}
\alias{extract_points}
\title{extract_points}
\usage{
extract_points(raster, pattern)
}
\arguments{
\item{raster}{RasterLayer.}

\item{pattern}{Point pattern.}
}
\value{
data.frame
}
\description{
Extract points
}
\details{
The function extracts the number of points within each habitat.
}
\examples{
landscape_classified <- classify_habitats(landscape, classes = 5)
extract_points(raster = landscape_classified, pattern = species_b)

}
