% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gest_multi.R
\name{Gest.Multi}
\alias{Gest.Multi}
\title{Multivariate G-function}
\usage{
Gest.Multi(pattern, r_max, r_length = 515)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr ppp object of the spatstat packages containing multivariate point pattern}

\item{r_max}{[\code{numeric(1)}]\cr Maximum distance r of multivariate G function}

\item{r_length}{[\code{numeric(1)}]\cr Number of evaluated distances r}
}
\value{
Data frame containing r and the multivaria G function for all species separately and the overall mean
}
\description{
The multivariate nearest neighbour distribution function G(r) to all points belonging to a different species than the observed point
}
