% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcf_fast.R
\name{Pcf.Fast}
\alias{Pcf.Fast}
\title{pcf-function}
\usage{
Pcf.Fast(pattern, correction = "good", method = "d", ...)
}
\arguments{
\item{pattern}{[\code{ppp(1)}] ppp object of the spatstat packages}

\item{correction}{[\code{string(1)}] Edge correction (see ?spatstat::Kest() for more details)}
}
\value{
fv object of the spatstat package
}
\description{
Computational fast method to calculate the pair correlation function using Ripleys K-Function
}
