% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rd_mar.R
\name{plot.rd_mar}
\alias{plot.rd_mar}
\title{plot.rd_mar}
\usage{
\method{plot}{rd_mar}(
  x,
  what = "sf",
  n = NULL,
  probs = c(0.025, 0.975),
  ask = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{rd_mar object with randomized patterns.}

\item{what}{Character specifying to plot summary functions of point patterns
(\code{what = "sf"}) or actual patterns (\code{what = "pp"}).}

\item{n}{Integer with number or vector of ids of randomized pattern to plot.
See Details section for more information.}

\item{probs}{Vector with quantiles of randomized data used for envelope construction.}

\item{ask}{Logical if the user is asked to press <RETURN> before second summary function
is plotted (only used if \code{what = "sf"}).}

\item{verbose}{Logical if progress report is printed.}

\item{...}{Not used.}
}
\value{
void
}
\description{
Plot method for rd_pat object
}
\details{
The function plots the pair correlation function and the nearest neighbour function of
the observed pattern and the reconstructed patterns (as "simulation envelopes").

It is also possible to plot n randomized patterns and the observed pattern
using \code{what = "pp"}. If \code{n} is a single number, \code{n} randomized
patterns will be sampled to plot. If \code{n} is a vector, the corresponding patterns
will be plotted.
}
\examples{
\dontrun{
pattern_recon <- reconstruct_pattern(species_a, n_random = 1, max_runs = 1000,
simplify = TRUE, return_input = FALSE)
marks_sub <- spatstat.geom::subset.ppp(species_a, select = dbh)
marks_recon <- reconstruct_pattern_marks(pattern_recon, marks_sub,
n_random = 19, max_runs = 1000)
plot(marks_recon)
}


}
\seealso{
\code{\link{reconstruct_pattern}} \cr
\code{\link{fit_point_process}}
}
