% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_randomized_pattern.R
\name{plot_randomized_pattern}
\alias{plot_randomized_pattern}
\title{plot_randomized_pattern}
\usage{
plot_randomized_pattern(pattern, comp_fast = FALSE, base_size = 15,
  size = 0.5)
}
\arguments{
\item{pattern}{List with reconstructed patterns}

\item{comp_fast}{Logical if summary functions should be estimated in an computational
fast way (but without edge correction)}

\item{base_size}{base font size}

\item{size}{Line size}
}
\value{
ggplot
}
\description{
Plot randomized pattern
}
\details{
Plot the summary functions of the observed pattern and the reconstructed patterns
as "simulation envelopes".
}
\examples{
\dontrun{
pattern_random <- spatstat::runifpoint(n = 50)
pattern_recon <- SHAR::reconstruct_pattern(pattern_random, n_random = 9, max_runs = 1000)
plot_randomized_pattern(pattern_recon)
}

}
