% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_randomized_pattern.R
\name{plot_randomized_pattern}
\alias{plot_randomized_pattern}
\title{plot_randomized_pattern}
\usage{
plot_randomized_pattern(pattern, size = 0.5, base_size = 15,
  comp_fast = FALSE)
}
\arguments{
\item{pattern}{List with reconstructed patterns.}

\item{size}{Line size.}

\item{base_size}{base font size.}

\item{comp_fast}{Should summary functions be estimated in an computational fast way.}
}
\value{
ggplot
}
\description{
Plot randomized pattern
}
\details{
The function plots the pair correlation function and the nearest neighbour function
the observed pattern and the reconstructed patterns (as "simulation envelopes".).
For large patterns `comp_fast = TRUE` decreases the computational demand because no edge
correction is used and the pair correlation function is estimated based on Ripley's
K-function. For more information see \code{\link{estimate_pcf_fast}}.
}
\examples{
\dontrun{
pattern_recon <- SHAR::reconstruct_pattern(species_a, n_random = 9, max_runs = 1000)
plot_randomized_pattern(pattern_recon)
}

}
