% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization_algorithm.R
\name{randomization_algorithm}
\alias{randomization_algorithm}
\title{Internal help function}
\usage{
randomization_algorithm(raster, number_neighbours = 8)
}
\arguments{
\item{raster}{[\code{raster(1)}] Raster of the raster object with discrete habitats}

\item{number_neighbours}{[\code{numeric(1)}] Number of random neigbhours. See ?raster::adjacent() for more details}
}
\value{
Raster of the raster package with randomized habitats
}
\description{
Randomization algorithm assigning habitat to random neighbours
}
