% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_pattern.R
\name{reconstruct_pattern}
\alias{reconstruct_pattern}
\title{reconstruct_pattern}
\usage{
reconstruct_pattern(pattern, n_random = 19, e_threshold = 0.01,
  max_runs = 10000, fitting = FALSE, comp_fast = FALSE,
  return_input = TRUE, verbose = FALSE, plot = FALSE)
}
\arguments{
\item{pattern}{List with reconstructed patterns.}

\item{n_random}{Number of randomized RasterLayers.}

\item{e_threshold}{Minimum energy to stop reconstruction.}

\item{max_runs}{Maximum number of iterations of e_threshold is not reached.}

\item{fitting}{It true, the pattern reconstruction starts with a fitting of a Thomas process.}

\item{comp_fast}{Should summary functions be estimated in an computational fast way.}

\item{return_input}{The original input data is returned as last list entry}

\item{verbose}{Print progress report.}

\item{plot}{Plot pcf function during optimization}
}
\value{
list
}
\description{
Pattern reconstruction
}
\details{
The functions randomizes the observed pattern by using pattern reconstruction
as described in Tscheschel & Stoyan (2006) and Wiegand & Moloney (2014). The
algorithm starts with a random reconstructed pattern, shifts a point to a new location and
keeps the change only, if the deviation between the observed and the reconstructed
pattern decreases. The pair correlation function and the nearest neighbour
distance function are used to describe the patterns. For large patterns
`comp_fast = TRUE` decreases the computational demand because no edge
correction is used and the pair correlation function is estimated based on Ripley's
K-function. For more information see \code{\link{estimate_pcf_fast}}.
}
\examples{
\dontrun{
pattern_recon <- reconstruct_pattern(species_b, n_random = 39, max_runs = 1000)
}

}
\references{
Tscheschel, A., & Stoyan, D. (2006). Statistical reconstruction of random point
patterns. Computational Statistics and Data Analysis, 51(2), 859–871.

Wiegand, T., & Moloney, K. A. (2014). Handbook of spatial point-pattern analysis
in ecology. Boca Raton: Chapman and Hall/CRC Press.
}
\seealso{
\code{\link{calculate_energy}} \cr
\code{\link{plot_randomized_pattern}}
}
