% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_habitat_association.R
\name{results_habitat_association}
\alias{results_habitat_association}
\title{results_habitat_association}
\usage{
results_habitat_association(pattern, raster, threshold = c(0.025, 0.975))
}
\arguments{
\item{pattern}{Point pattern or list with reconstructed patterns}

\item{raster}{RasterLayer or list of RasterLayers}

\item{threshold}{Significance thresholds}
}
\value{
list
}
\description{
Results habitat association
}
\details{
Results of habitat association tests
}
\examples{
\dontrun{
pattern_random <- spatstat::runifpoint(n = 50)
pattern_recon <- SHAR::reconstruct_pattern(pattern_random, n_random = 9, max_runs = 1000)
}

}
\references{
Harms, K. E., Condit, R., Hubbell, S. P., & Foster, R. B. (2001). Habitat associations
of trees and shrubs in a 50-ha neotropical forest plot. Journal of Ecology, 89(6), 947–959.

Plotkin, J. B., Potts, M. D., Leslie, N., Manokaran, N., LaFrankie, J. V., & Ashton, P. S. (2000).
Species-area curves, spatial aggregation, and habitat specialization in tropical forests.
Journal of Theoretical Biology, 207(1), 81–99.
}
\seealso{
\code{\link{calculate_mean_energy}} \cr
\code{\link{plot_randomized_pattern}}
}
