% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_habitat_association.R
\name{results_habitat_association}
\alias{results_habitat_association}
\title{results_habitat_association}
\usage{
results_habitat_association(
  pattern,
  raster,
  significance_level = 0.05,
  breaks = NULL,
  digits = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pattern}{ppp object with original point pattern data or rd_pat or rd_mar
object with randomized point pattern.}

\item{raster}{SpatRaster with original discrete habitat data or rd_ras object with
randomized environmental data.}

\item{significance_level}{Double with significance level.}

\item{breaks}{Vector with breaks of habitat classes.}

\item{digits}{Integer with digits used during rounding.}

\item{verbose}{Logical if messages should be printed.}
}
\value{
data.frame
}
\description{
Results habitat association
}
\details{
The functions shows significant habitat associations by comparing the number of
points within a habitat between the observed data and randomized data as described in
Plotkin et al. (2000) and Harms et al. (2001). Significant positive or associations are present
if the observed count in a habitat is above or below a certain threshold of the
randomized count, respectively.

In case the SpatRaster contains NA cells, this needs to be reflected in the observation
window of the point pattern as well (i.e., no point locations possible in these areas).

If \code{breaks = NULL} (default), only habitat labels (but not breaks) will be
returned. If a vector with \code{breaks} is provided (same order as increasing habitat values),
the breaks will be included as well.
}
\examples{
landscape_classified <- classify_habitats(terra::rast(landscape), n = 5, style = "fisher")
species_a_random <- fit_point_process(species_a, n_random = 199)
results_habitat_association(pattern = species_a_random, raster = landscape_classified)

}
\references{
Harms, K.E., Condit, R., Hubbell, S.P., Foster, R.B., 2001. Habitat associations of
trees and shrubs in a 50-ha neotropical forest plot. Journal of Ecology 89, 947–959.
<https://doi.org/10.1111/j.1365-2745.2001.00615.x>

Plotkin, J.B., Potts, M.D., Leslie, N., Manokaran, N., LaFrankie, J.V.,
Ashton, P.S., 2000. Species-area curves, spatial aggregation, and habitat specialization
in tropical forests. Journal of Theoretical Biology 207, 81–99.
<https://doi.org/10.1006/jtbi.2000.2158>
}
\seealso{
\code{\link{reconstruct_pattern}},
\code{\link{fit_point_process}}
}
