% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_habitat_association.R
\name{results_habitat_association}
\alias{results_habitat_association}
\title{results_habitat_association}
\usage{
results_habitat_association(pattern, raster, threshold = c(0.025, 0.975))
}
\arguments{
\item{pattern}{Point pattern or list with reconstructed patterns.}

\item{raster}{RasterLayer or list of RasterLayers.}

\item{threshold}{Significance thresholds.}
}
\value{
data.frame
}
\description{
Results habitat association
}
\details{
The functions shows significant habitat associations by comparing the number of
points within a habitat between the observed data and randomized data as described in
Plotkin et al. (2000) and Harms et al. (2001). Significant positive or associations are present
if the observed count in a habitat is above or below a certain threshold of the
randomized count, respectively.
}
\examples{
\dontrun{
landscape <- NLMR::nlm_fbm(ncol = 50, nrow = 50, user_seed = 1)
landscape_classified <- SHAR::classify_habitats(landscape, classes = 5)
species_1 <- spatstat::runifpoint(n = 50, win = spatstat::owin(c(0, 50), c(0, 50)))
species_1_reconstructed <- SHAR::reconstruct_pattern(pattern = species_1,
n_random = 9, max_runs = 1000)
results_habitat_association(pattern = species_1_reconstructed, raster = landscape_classified)
}

}
\references{
Harms, K. E., Condit, R., Hubbell, S. P., & Foster, R. B. (2001). Habitat associations
of trees and shrubs in a 50-ha neotropical forest plot. Journal of Ecology, 89(6), 947–959.

Plotkin, J. B., Potts, M. D., Leslie, N., Manokaran, N., LaFrankie, J. V., & Ashton, P. S. (2000).
Species-area curves, spatial aggregation, and habitat specialization in tropical forests.
Journal of Theoretical Biology, 207(1), 81–99.
}
\seealso{
\code{\link{randomization_algorithm}} \cr
\code{\link{reconstruct_pattern}}
}
