% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAR.R
\docType{package}
\name{SHAR-package}
\alias{SHAR}
\alias{SHAR-package}
\title{SHAR}
\description{
\emph{SHAR} provides ,,,
}
\author{
\strong{Maintainer}: Maximillian H.K. Hesselbarth \email{maximilian.hesselbarth@uni-goettingen.de} (0000-0003-1125-9918)

Authors:
\itemize{
  \item Marco Sciaini \email{sciaini.marco@gmail.com} (0000-0002-3042-5435)
}

}
