% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAR.R
\docType{package}
\name{shar}
\alias{shar}
\alias{shar-package}
\title{shar}
\description{
Analyse species-habitat associations in R. Therefore, information about the
location of the species is needed and about the environmental conditions. To test
for significance habitat associations, one of the two components is randomized.
Methods are mainly based on Plotkin et al. (2000) <doi:10.1006/jtbi.2000.2158> and
Harms et al. (2001) <doi:10.1111/j.1365-2745.2001.00615.x>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-spatialecology.github.io/shar}
  \item Report bugs at \url{https://github.com/r-spatialecology/shar/issues}
}

}
\author{
\strong{Maintainer}: Maximillian H.K. Hesselbarth \email{maximilian.hesselbarth@uni-goettingen.de} (0000-0003-1125-9918)

Authors:
\itemize{
  \item Marco Sciaini \email{sciaini.marco@gmail.com} (0000-0002-3042-5435)
}

}
