% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_reconstruction.R
\name{Spatial.Reconstruction}
\alias{Spatial.Reconstruction}
\title{Spatial reconstruction}
\usage{
Spatial.Reconstruction(pattern, max_runs = 10000, e_threshold = 0.01,
  fitting = F, verbose = T)
}
\arguments{
\item{pattern}{[\code{ppp(1)}] Multivariate ppp object of the spatstat package}

\item{max_runs}{[\code{numeric(1)}] number of maximum iterations}

\item{e_threshold}{[\code{numeric(1)}] Threshold for energy to reach during reconstruction}

\item{fitting}{[\code{logical(1)}] If TRUE, a clustered pattern is fitted to the original pattern as starting point}

\item{verbose}{[\code{logical(1)}]\cr If TRUE, progress is printed}
}
\value{
ppp object of the spatstat package with reconstructed pattern
}
\description{
Univariate pattern reconstruction of only the spatial structure
}
