% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_a}
\alias{species_a}
\title{Species a}
\format{A spatstat ppp object.}
\usage{
species_a
}
\description{
A species with negative associations to habitat 4 of `landscape`.
}
\keyword{datasets}
