% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_b}
\alias{species_b}
\title{Species b}
\format{A spatstat ppp object.}
\usage{
species_b
}
\description{
A species with positive associations to habitat 5 of `landscape`.
}
\keyword{datasets}
